restructuring application sidebar
