+++ clients/core/src/PhaseMapping/PhaseSidebarMapping.tsx
@@ -1,41 +1,6 @@
-import { Role } from '@/interfaces/permission_roles'
 import { TemplateSidebar } from './ExternalSidebars/TemplateSidebar'
-import { FileUser } from 'lucide-react'
-import { ExternalSidebarComponent } from './ExternalSidebars/ExternalSidebar'
-import { SidebarMenuItemProps } from '@/interfaces/sidebar'
 import { InterviewSidebar } from './ExternalSidebars/InterviewSidebar'
-
-const ApplicationSidebar = ({ rootPath, title }: { rootPath: string; title: string }) => {
-  const applicationSidebarItems: SidebarMenuItemProps = {
-    title: 'Application',
-    icon: <FileUser />,
-    goToPath: '',
-    subitems: [
-      {
-        title: 'Applications',
-        goToPath: '/applications',
-        requiredPermissions: [Role.PROMPT_ADMIN, Role.COURSE_LECTURER],
-      },
-      {
-        title: 'Configuration',
-        goToPath: '/configuration',
-        requiredPermissions: [Role.PROMPT_ADMIN, Role.COURSE_LECTURER],
-      },
-      {
-        title: 'Mailing',
-        goToPath: '/mailing',
-        requiredPermissions: [Role.PROMPT_ADMIN, Role.COURSE_LECTURER],
-      },
-    ],
-  }
-  return (
-    <ExternalSidebarComponent
-      title={title}
-      rootPath={rootPath}
-      sidebarElement={applicationSidebarItems}
-    />
-  )
-}
+import { ApplicationSidebar } from './ExternalSidebars/ApplicationSidebar'
 
 export const PhaseSidebarMapping: { [key: string]: React.FC<{ rootPath: string; title: string }> } =
   {
